requirements.txt
pytube==11.0.1
requests==2.26.0
git+https://github.com/Uberi/speech_recognition@master#egg=speechrecognition
wheel==0.37.0
google-api-python-client==2.31.0
google-cloud-speech==2.11.1
oauth2client==4.1.3
python-dotenv==0.19.2
